# api/gateway/analysis/api.yml
...
swagger: '2.0'
info:
  version: '1.0.0' # API version, not the SAR code version
  title: panther-analysis-api
  description: Panther Python policies/rules
  contact:
    name: Panther Labs
    url: https://runpanther.io/about
    email: support@runpanther.io
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json

# The name of the Lambda handler function that will process API requests
...
paths:
...
  /list:
    # The frontend pages through the policies in a customer account.
    #
    # Example: GET /list ?
    #     complianceStatus=FAIL &
    #     nameContains=Encryption &
    #     enabled=true &
    #     hasRemediation=true &
    #     resourceTypes=AWS.S3.Bucket,AWS.KMS.Key
    #     severity=MEDIUM &
    #     tags=aws,panther &
    #     sortBy=lastModified
    #     sortDir=ascending
    #
    # Response: {
    #     "paging": {
    #         "thisPage":   1,
    #         "totalPages": 1,
    #         "totalItems": 1
    #     },
    #     "policies": [
    #         {
    #             "autoRemediationId": "some-fix",
    #             "complianceStatus":  "FAIL",
    #             "displayName":       "",
    #             "enabled":           true,
    #             "id":                "AWS.S3.BucketEncryption",
    #             "lastModified":      "2019-08-26T00:00:00.000Z",
    #             "resourceTypes":     ["AWS.KMS.Key", "AWS.S3.Bucket"],
    #             "severity":          "MEDIUM",
    #             "tags":              ["aws", "s3", "cis"]
    #         }
    #     ]
    # }
    get:
      operationId: ListPolicies
      summary: Page through policies in a customer's account
      parameters:
        # filtering
        - name: complianceStatus
          in: query
          description: Only include policies with a specific compliance status
          type: string
          enum: [ERROR, FAIL, PASS]
        - name: nameContains
          in: query
          description: Only include policies whose ID or display name contains this substring (case-insensitive)
          type: string
        - name: enabled
          in: query
          description: Only include policies which are enabled or disabled
          type: boolean
        - name: hasRemediation
          in: query
          description: Only include policies with or without auto-remediation enabled
          type: boolean
        - name: resourceTypes
          in: query
          description: Only include policies which apply to one of these resource types
          type: array
          collectionFormat: csv
          uniqueItems: true
          items:
            type: string
        - name: severity
          in: query
          description: Only include policies with this severity
          type: string
          enum: [INFO, LOW, MEDIUM, HIGH, CRITICAL]
        - name: tags
          in: query
          description: Only include policies with all of these tags (case-insensitive)
          type: array
          collectionFormat: csv
          uniqueItems: true
          items:
            type: string

        # sorting
        - name: sortBy
          in: query
          description: Name of the field to sort by
          type: string
          enum:
            - complianceStatus
            - enabled
            - id
            - lastModified
            - resourceTypes
            - severity
          default: severity
        - name: sortDir
          in: query
          description: Sort direction
          type: string
          enum: [ascending, descending]
          default: descending

        # paging
        - name: pageSize
          in: query
          description: Number of items in each page of results
          type: integer
          minimum: 1
          maximum: 1000
          default: 25
        - name: page
          in: query
          description: Which page of results to retrieve
          type: integer
          minimum: 1
          default: 1
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/PolicyList'
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error
...
  /rule/list:
    # Same as ListPolicies, but for log analysis rules
    get:
      operationId: ListRules
      summary: Page through rules in a customer's account
      parameters:
        # filtering
        - name: nameContains
          in: query
          description: Only include rules whose ID or display name contains this substring (case-insensitive)
          type: string
        - name: enabled
          in: query
          description: Only include rules which are enabled or disabled
          type: boolean
        - name: logTypes
          in: query
          description: Only include rules which apply to one of these log types
          type: array
          collectionFormat: csv
          uniqueItems: true
          items:
            type: string
        - name: severity
          in: query
          description: Only include policies with this severity
          type: string
          enum: [INFO, LOW, MEDIUM, HIGH, CRITICAL]
        - name: tags
          in: query
          description: Only include policies with all of these tags (case-insensitive)
          type: array
          collectionFormat: csv
          uniqueItems: true
          items:
            type: string

        # sorting
        - name: sortBy
          in: query
          description: Name of the field to sort by
          type: string
          enum:
            - enabled
            - id
            - lastModified
            - logTypes
            - severity
          default: severity
        - name: sortDir
          in: query
          description: Sort direction
          type: string
          enum: [ascending, descending]
          default: ascending

        # paging
        - name: pageSize
          in: query
          description: Number of items in each page of results
          type: integer
          minimum: 1
          maximum: 1000
          default: 25
        - name: page
          in: query
          description: Which page of results to retrieve
          type: integer
          minimum: 1
          default: 1
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/RuleList'
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error
...
definitions:
  Error:
    type: object
    properties:
      message:
        description: Error message
        type: string
    required:
      - message

  ##### Create/Update/Modify #####
...
  TypeSet:
    description: List of resource/log types to which this policy applies
    type: array
    maxItems: 500
    uniqueItems: true
    items:
      type: string
      minLength: 1
...
  PolicyList:
    type: object
    properties:
      paging:
        $ref: '#/definitions/Paging'
      policies:
        type: array
        items:
          $ref: '#/definitions/PolicySummary'
    required:
      - paging
      - policies
  Paging:
    type: object
    properties:
      thisPage:
        type: integer
        minimum: 1
      totalPages:
        type: integer
        minimum: 1
      totalItems:
        type: integer
        minimum: 1
    required:
      - thisPage
      - totalPages
      - totalItems
  PolicySummary:
    type: object
    properties: # only the fields we need for the table in the UI
      autoRemediationId:
        $ref: '#/definitions/autoRemediationId'
      autoRemediationParameters:
        $ref: '#/definitions/autoRemediationParameters'
      complianceStatus:
        $ref: '#/definitions/complianceStatus'
      displayName:
        $ref: '#/definitions/displayName'
      enabled:
        $ref: '#/definitions/enabled'
      id:
        $ref: '#/definitions/id'
      lastModified:
        $ref: '#/definitions/modifyTime'
      resourceTypes:
        $ref: '#/definitions/TypeSet'
      severity:
        $ref: '#/definitions/severity'
      tags:
        $ref: '#/definitions/tags'
      suppressions:
        $ref: '#/definitions/suppressions'
    required:
      - autoRemediationId
      - autoRemediationParameters
      - complianceStatus
      - displayName
      - enabled
      - id
      - lastModified
      - resourceTypes
      - severity
      - tags
      - suppressions

  ##### BulkUpload #####
...
  RuleList:
    type: object
    properties:
      paging:
        $ref: '#/definitions/Paging'
      rules:
        type: array
        items:
          $ref: '#/definitions/RuleSummary'
    required:
      - paging
      - rules
  RuleSummary:
    type: object
    properties: # only the fields we need for the table in the UI
      displayName:
        $ref: '#/definitions/displayName'
      enabled:
        $ref: '#/definitions/enabled'
      id:
        $ref: '#/definitions/id'
      lastModified:
        $ref: '#/definitions/modifyTime'
      logTypes:
        $ref: '#/definitions/TypeSet'
      severity:
        $ref: '#/definitions/severity'
      tags:
        $ref: '#/definitions/tags'
    required:
      - displayName
      - enabled
      - id
      - lastModified
      - logTypes
      - severity
      - tags

  ##### object properties #####
  autoRemediationId:
    description: When a resource fails the policy, trigger the remediation with this ID
    type: string
    maxLength: 200
  autoRemediationParameters:
    description: Configuration parameters passed to the remediation handler
    type: object
    additionalProperties:
      type: string
    maxProperties: 50
...
  complianceStatus:
    description: Pass/fail compliance status
    type: string
    enum:
      - ERROR # Policy raised an exception when evaluating a resource
      - FAIL # Policy failed on at least one resource
      - PASS # Policy passed for all applicable resources
...
  displayName:
    description: Optional user-friendly policy name
    type: string
    maxLength: 200
  enabled:
    description: True if the policy is currently being evaluated
    type: boolean
  id:
    description: User-specified unique rule/policy ID
    type: string
    pattern: '[a-zA-Z0-9\-\. ]{1,200}'
  modifyTime:
    description: Policy modification timestamp
    type: string
    format: date-time
...
  suppressions:
    description: >
      List of resource ID regexes that are excepted from this policy.
      The policy will still be evaluated, but failures will not trigger alerts nor remediations.
    type: array
    maxItems: 500
    uniqueItems: true
    items:
      type: string
      maxLength: 1000
...
  severity:
    description: Policy severity
    type: string
    enum:
      - INFO
      - LOW
      - MEDIUM
      - HIGH
      - CRITICAL
  tags:
    description: List of tags to categorize the policy
    type: array
    maxItems: 500
    uniqueItems: true
    items:
      type: string
      maxLength: 200
...